core/requirements.txt
alembic==0.8.6
dataset==0.6.4
Mako==1.0.4
MarkupSafe==0.23
normality==0.2.4
python-dateutil==2.5.3
python-editor==1.0
pytz==2016.4
PyYAML==3.11
requests==2.10.0
six==1.10.0
slackweb==1.0.5
SQLAlchemy==1.0.13
